+++ CHANGES.rst
@@ -4,13 +4,24 @@ Changelog
 3.0b5 (unreleased)
 ------------------
 
+Breaking changes
+++++++++++++++++
+
+- Removed ``VSEval`` module. Please use DT_Util.EVal now.
+
+Features
+++++++++
+
+- Added support for Python 3.7.
+
+Bugfixes
+++++++++
+
 - Fix regression with exception handling in ``<dtml-except>`` with Python 2.
   (`#25 <https://github.com/zopefoundation/DocumentTemplate/issues/25>`_)
 
 - Stabilized TreeTag rendering for objects without ``_p_oid`` values.
   (`#26 <https://github.com/zopefoundation/DocumentTemplate/issues/26>`_)
-
-- Added support for Python 3.7.
 
 - Remove support for string exceptions in ``<dtml-except>``.
   (`#29 <https://github.com/zopefoundation/DocumentTemplate/pull/29>`_)